move dist utils to `distributed` package
