+++ src/main/scala/org/hammerlab/guacamole/commands/GermlineThresholdCaller.scala
@@ -23,9 +23,10 @@ import org.apache.spark.rdd.RDD
 import org.bdgenomics.formats.avro.GenotypeAllele.{Alt, NoCall, OtherAlt, Ref}
 import org.bdgenomics.formats.avro.{Contig, Genotype, GenotypeAllele, Variant}
 import org.hammerlab.guacamole.Common.Arguments.GermlineCallerArgs
-import org.hammerlab.guacamole.loci.LociPartitionUtils.partitionLociAccordingToArgs
+import org.hammerlab.guacamole.distributed.{LociPartitionUtils, PileupFlatMapUtils}
+import LociPartitionUtils.partitionLociAccordingToArgs
 import org.hammerlab.guacamole.pileup.Pileup
-import org.hammerlab.guacamole.pileup.PileupFlatMapUtils.pileupFlatMap
+import PileupFlatMapUtils.pileupFlatMap
 import org.hammerlab.guacamole.reads.Read
 import org.hammerlab.guacamole.reference.ReferenceBroadcast
 import org.hammerlab.guacamole.variants.Allele
