+++ src/main/scala/org/hammerlab/guacamole/commands/VAFHistogram.scala
@@ -9,10 +9,11 @@ import org.apache.spark.mllib.linalg.Vectors
 import org.apache.spark.rdd.RDD
 import org.apache.spark.storage.StorageLevel
 import org.hammerlab.guacamole._
-import org.hammerlab.guacamole.loci.LociPartitionUtils.partitionLociAccordingToArgs
-import org.hammerlab.guacamole.loci.{LociMap, LociPartitionUtils}
+import org.hammerlab.guacamole.distributed.{LociPartitionUtils, PileupFlatMapUtils}
+import LociPartitionUtils.partitionLociAccordingToArgs
+import org.hammerlab.guacamole.loci.LociMap
 import org.hammerlab.guacamole.pileup.Pileup
-import org.hammerlab.guacamole.pileup.PileupFlatMapUtils.pileupFlatMap
+import PileupFlatMapUtils.pileupFlatMap
 import org.hammerlab.guacamole.reads.Read.InputFilters
 import org.hammerlab.guacamole.reads.{MappedRead, Read}
 import org.hammerlab.guacamole.reference.{ReferenceBroadcast, ReferenceGenome}
