+++ src/main/scala/org/hammerlab/guacamole/distributed/LociPartitionUtils.scala
@@ -1,9 +1,10 @@
-package org.hammerlab.guacamole.loci
+package org.hammerlab.guacamole.distributed
 
 import org.apache.spark.rdd.RDD
 import org.hammerlab.guacamole.Common.Arguments.{Base, Loci}
 import org.hammerlab.guacamole.Common._
 import org.hammerlab.guacamole.HasReferenceRegion
+import org.hammerlab.guacamole.loci.{LociMap, LociSet}
 import org.kohsuke.args4j.{Option => Args4jOption}
 
 import scala.reflect.ClassTag
@@ -34,7 +35,8 @@ object LociPartitionUtils {
         tasks,
         loci,
         args.partitioningAccuracy,
-        regionRDDs: _*)
+        regionRDDs: _*
+      )
     }
   }
