+++ src/main/scala/org/hammerlab/guacamole/distributed/PileupFlatmapUtils.scala
@@ -1,12 +1,13 @@
-package org.hammerlab.guacamole.pileup
+package org.hammerlab.guacamole.distributed
 
 import org.apache.spark.rdd.RDD
 import org.hammerlab.guacamole._
 import org.hammerlab.guacamole.loci.LociMap
+import org.hammerlab.guacamole.pileup.Pileup
 import org.hammerlab.guacamole.reads.MappedRead
 import org.hammerlab.guacamole.reference.{ReferenceGenome, _}
 import org.hammerlab.guacamole.windowing.SlidingWindow
-import org.hammerlab.guacamole.windowing.WindowFlatMapUtils.windowFlatMapWithState
+import WindowFlatMapUtils.windowFlatMapWithState
 
 import scala.reflect.ClassTag
