+++ src/main/scala/org/hammerlab/guacamole/distributed/RegionsByContig.scala
@@ -1,4 +1,4 @@
-package org.hammerlab.guacamole.util
+package org.hammerlab.guacamole.distributed
 
 import org.hammerlab.guacamole.HasReferenceRegion
