+++ src/main/scala/org/hammerlab/guacamole/distributed/TaskPosition.scala
@@ -1,4 +1,4 @@
-package org.hammerlab.guacamole.util
+package org.hammerlab.guacamole.distributed
 
 /**
   * TaskPosition represents the task a read is assigned to and the start position on the reference genome of the read
