+++ src/main/scala/org/hammerlab/guacamole/distributed/WindowFlatMapUtils.scala
@@ -1,11 +1,11 @@
-package org.hammerlab.guacamole.windowing
+package org.hammerlab.guacamole.distributed
 
 import org.apache.commons.math3
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.rdd.RDD
 import org.hammerlab.guacamole.Common._
 import org.hammerlab.guacamole.loci.{LociMap, LociSet}
-import org.hammerlab.guacamole.util.{HashMapAccumulatorParam, KeyPartitioner, RegionsByContig, TaskPosition}
+import org.hammerlab.guacamole.windowing.{SlidingWindow, SplitIterator}
 import org.hammerlab.guacamole.{DelayedMessages, HasReferenceRegion, PerSample}
 
 import scala.collection.mutable.{HashMap => MutableHashMap}
@@ -30,6 +30,7 @@ object WindowFlatMapUtils {
     *                       then it is included.
     * @param initialState initial state to use for each task and each contig analyzed within a task.
     * @param function function to flatmap, of type (state, sliding windows) -> (new state, result data)
+    * @tparam M region data type (e.g. MappedRead)
     * @tparam T result data type
     * @tparam S state type
     * @return RDD[T] of flatmap results
@@ -132,7 +133,7 @@ object WindowFlatMapUtils {
     * @tparam T type of value returned by function
     * @return flatMap results, RDD[T]
     */
-  private[windowing] def windowTaskFlatMapMultipleRDDs[M <: HasReferenceRegion: ClassTag, T: ClassTag](
+  private[distributed] def windowTaskFlatMapMultipleRDDs[M <: HasReferenceRegion: ClassTag, T: ClassTag](
     regionRDDs: PerSample[RDD[M]],
     lociPartitions: LociMap[Long],
     halfWindowSize: Long,
